# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Staffio HR Management API
  version: 1.0.0
  description: |
    API do sistema de gestão de RH (Staffio). Backend em Go (Gin).

    Autenticação é baseada em sessão (cookie). A maioria dos endpoints exige
    sessão autenticada e os endpoints de escrita respeitam a hierarquia de papéis:
    - `owner`   (dono)  – acesso total
    - `admin`   (admin) – gerencia equipe, edita candidatos
    - `recruit` (recrutador) – edita candidatos e departamentos, sem relatórios/visão geral
    - `viewer`  (visualizador) – somente leitura; bloqueado para métodos não-GET

    Vários endpoints renderizam páginas HTML (servidor). Apenas os endpoints
    prefixados com `/api/` retornam JSON estruturado.
  contact:
    name: Staffio
servers:
  - url: http://localhost:8000
    description: Servidor local
...
paths:
...
  /login:
    get:
      tags: [Pages, Auth]
      summary: Página de login
      description: Renderiza o formulário de login (HTML).
      responses:
        '200':
          description: Página de login renderizada
          content:
            text/html:
              schema: { type: string }
    post:
      tags: [Auth]
      summary: Autenticar usuário
      description: Valida credenciais e cria uma sessão (cookie). Limitado a 10 requisições por minuto por IP.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [username, email, password]
              properties:
                username:
                  type: string
                  description: Nome de usuário
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '302':
          description: Redireciona para /dashboard em caso de sucesso
        '400':
          description: Campos obrigatórios ausentes
        '401':
          description: Credenciais incorretas
...
  /employees:
    get:
      tags: [Pages, Employees]
      summary: Listar candidatos (página)
      description: Página com listagem paginada e contagens por status. Requer autenticação.
      security:
        - sessionAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema: { type: integer, default: 1, minimum: 1 }
        - name: status
          in: query
          required: false
          description: Filtro de status (interviewing = pending, hired = contractors)
          schema:
            type: string
            enum: [all, interviewing, hired, rejected]
            default: all
      responses:
        '200':
          description: Página de candidatos renderizada
          content:
            text/html: { schema: { type: string } }
        '302':
          description: Redireciona para /login se não autenticado
    post:
      tags: [Employees]
      summary: Criar candidato
      description: Cria um novo candidato (status inicial pending). Requer papel com permissão de edição (owner, admin, recruit).
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [full_name, email, position]
              properties:
                full_name:
                  type: string
                email:
                  type: string
                  format: email
                position:
                  type: string
                description:
                  type: string
                photo:
                  type: string
                  format: binary
                  description: Foto opcional (JPG, PNG, GIF, WebP, máx. 5 MB)
      responses:
        '302':
          description: Redireciona para /employees em caso de sucesso
        '400':
          description: Campos obrigatórios ausentes ou upload inválido
        '401':
          description: Não autenticado
        '403':
          description: Sem permissão de edição
        '500':
          description: Erro ao salvar no banco
...
  /department:
    get:
      tags: [Pages, Departments]
      summary: Página de departamentos
      description: Lista departamentos e funcionários. Recrutadores são redirecionados para /employees.
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Página de departamentos renderizada
          content:
            text/html: { schema: { type: string } }
        '302':
          description: Redireciona se não autenticado ou se é recrutador
        '500':
          description: Erro ao buscar dados
    post:
      tags: [Departments]
      summary: Criar departamento
      description: Cria um novo departamento. Requer permissão de gestão de equipe (owner, admin).
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [code, name]
              properties:
                code:
                  type: string
                name:
                  type: string
                boss_id:
                  type: integer
                  description: ID do gerente (opcional)
      responses:
        '302':
          description: Redireciona para /department em caso de sucesso
        '400':
          description: Campos obrigatórios ausentes ou ID de gerente inválido
        '401':
          description: Não autenticado
        '403':
          description: Sem permissão
        '500':
          description: Erro ao salvar departamento
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
      description: Cookie de sessão definido pelo servidor após login
...